Added parallax background and rain

+++ include/Iohandler.h
@@ -29,9 +29,9 @@ public:
 
     void windowtick();
 
-    void drawFromSheet(IntRect sheetrect, int x, int y, int w = 1, int h = 1, int t = 255);
+    void drawFromSheet(IntRect sheetrect, int x, int y, double w = 1, double h = 1, int t = 255);
 
-    void drawAddress(const string& address, int x, int y, int w = 1, int h = 1);
+    void drawAddress(const string& address, int x, int y, double w = 1, double h = 1);
 
 };
